Code refactoring

+++ Source/KeyPressingManager.cpp
@@ -13,7 +13,7 @@ void KeyPressingManager::readClickedKeys()
 {
     const unsigned keyAmt = Settings::LogicalKeys.size();
     const unsigned btnAmt = Settings::LogicalButtons.size();
-    size_t i = 0;
+    unsigned i = 0;
     for (; i < keyAmt; i++)
     {
         if (sf::Keyboard::isKeyPressed(Settings::LogicalKeys[i]->key) && !mNeedToBeReleased[i])
@@ -24,7 +24,7 @@ void KeyPressingManager::readClickedKeys()
         else
             mNeedToBeReleased[i] = false;
     }
-    for (size_t j = 0; j < btnAmt; j++, i++)
+    for (unsigned j = 0; j < btnAmt; j++, i++)
     {
         if (sf::Mouse::isButtonPressed(Settings::LogicalButtons[j]->button) && !mNeedToBeReleased[i])
             ++mClickedKeys[i];
